Made the functions compatible on my computer. Generalized data loading using pandas instead of numpy to make it more flexible for datasets that aren't all floats, made it possible to run with default settings for numbers of pc's, and separated out the functions into a separate file for better organization and easier troubleshooting.
